move type casting to the paymob class

+++ config/payment.php
@@ -1,9 +1,9 @@
 <?php
 
 return [
-    'paymob' => [
-        'api_key' => env('PAYMOB_API_KEY', ''),
-        'zeal_auth_integration' => (int) env('PAYMOB_AUTH_INTEGRATION', ''),
-        'zeal_payment_integration' => (int) env('PAYMOB_PAYMENT_INTEGRATION', ''),
-    ],
+		'paymob' => [
+				'api_key' => env('PAYMOB_API_KEY', ''),
+				'zeal_auth_integration' => env('PAYMOB_AUTH_INTEGRATION', ''),
+				'zeal_payment_integration' => env('PAYMOB_PAYMENT_INTEGRATION', ''),
+		],
 ];
